支付完成增加 payssion支付 增加mol支付测试

+++ luckybuyer/app/src/main/java/net/iwantbuyer/bean/AllOrderBean.java
@@ -468,7 +468,7 @@ public class AllOrderBean implements Serializable{
                 this.lucky_user = lucky_user;
             }
 
-            public int getNum_orders() {
+            public double getNum_orders() {
                 return num_orders;
             }
